// LoginForms/theme/types/IThemeTemplate.ts
export type ThemeTemplate = {
  themeColor: string;
  button: ThemeButton;
  text: ThemeText;
  textInput: ThemeTextInput;
  textInputUnderline: ThemeTextInputUnderline;
  dropdown: ThemeDropdown;
  captcha: ThemeCaptcha;
  radioButton: ThemeRadioButton;
  imageButton: ThemeImageButton;
  separator: ThemeSeparator;
  bookingSlots: ThemeBookingSlots;
  amenitiesList: ThemeAmenitiesList;
  switchButton: ThemeSwitchButton;
  tab: ThemeTab;
  hhoCompanyHeader: ThemeHHOCompanyHeader;
  moveInHeader: ThemeMoveInHeader;
  addFile: ThemeAddFile;
  addedFile: ThemeAddedFile;
  bottomSheet: ThemeBottomSheet;
  buttonBack: ThemeButtonBack;
  registerMoveInOption: ThemeMoveInRegisterOption;
  tableCell: ThemeTableCell;
  switch: ThemeSwitch;
  catalogueText: ThemeCatalogueText;
  addressText: ThemeAddressText;
  communityHeaderText: ThemeCommunityHeaderText;
  communityText: ThemeCommunityText;
  communityApartmentText: ThemeApartmentText;
  carosuel: ThemeCarosuel;
  menuButton: ThemeMenuBotton;
  textInputOutline: ThemeTextInputOutline;
  switchSelector: ThemeSwitchSelector;
  amenitiesSummery: ThemeAmenitiesSummery;
  amenitiesSuccess: ThemeAmenitiesSuccess;
  view: ThemeView;
  accessCard: ThemeAccessCard;
  propertyList: ThemePropertyList;
  serviceRequest: ThemeServiceRequest;
  lineSeparator:ThemeLineSeparator;
  moveOut:ThemeMoveOut;
  dashboard:ThemeDashboard;
  fileBrowser: ThemeFileBrowser;
  checkBox:ThemeCheckBox;
  loginTheme :ThemeLogin;
  alert: ThemeAlert;
  documentRequired: ThemeDocumentRequired;
  buying: ThemeBuying;
  moveInSuccess: ThemeMoveInSuccessPopup;
  moveInRequestItem: ThemeMoveInRequestItem;
  moveOutRequestItem: ThemeMoveOutRequestItem
};